requirements.txt
sanic
PyGithub
cachetools
python-dotenv
